Optimize execution time for listing separate obs
Substantially reduce the execution time by using a specific function
instead of comparing each unique observation to the whole set of
observations to compute the total weight of each unique observation.

Examples of improvement (time for step 2):
 - With 18.658 observation-year combinations: [06m 26s] -> [00s]
 - With 15.923 observation-year combinations: [03m 26s] -> [00s]
 - With 19.450 observation-year combinations: [11m 13s] -> [00s]

Nodes are now sorted in alphanumeric order of the items they contain
for nodes of the same length and weight.