Refactor one attribute of SA for consistency
Modification of the attribute node_per_year. It was a data.frame of
columns "node", "weight" and "year". It now has the same structure as
the attribute patterns_per_year, avoiding duplication of data and using
a more efficient structure.

The internal (Roxygen) and external (vignettes) documentation as well
as the functions that used the attribute have changed accordingly.

We can note a slight decrease in the total weight of the final object:
 - With a set of 6.389 nodes: 415.9 MB -> 412.3 MB
 - With a set of 3.952 nodes: 293.3 MB -> 236 MB
 - With a set of 9.025 nodes: 814.7 MB -> 811.2 MB
As well as slight improvements in time:
 - With 6.389 nodes: step 1: [07s] -> [04s]; step 7: [12s] -> [00s]
 - With 3.952 nodes: step 1: [07s] -> [05s]; step 7: [12s] -> [00s]
 - With 9.025 nodes: step 1: [07s] -> [02s]; step 7: [22s] -> [00s]