Reduce the time to build a spectrosome chart
Reduce the execution time to build a spectrosome chart in the case
where it is to be built on a subset of nodes or patterns.
More specifically, the optimization is in the search for the subset of
links associated with this subset of nodes or patterns.

Examples of improvement (function spectrosome_chart):
 - Plot of 10 nodes out of 6,340 which involve the extraction of 45
   links out of 6,858,520: [01m 31s] -> [05s]
 - Plot of 100 nodes out of 6,340 which involve the extraction of 4,176
   links out of 6,858,520: [02m 00s] -> [11s]
In the second example, the optimization leads to a time of 5 seconds
for the search for the subset of links. The rest of the time
corresponds to the plot of the graph.