fix: Correct the indexes of new isolated vertices
Ensure that vertices that become isolated due to the use of a subset
have a different index from those existing in the data frame of the all
set of links.
This happens when searching for links from a subset of nodes or
patterns, or when creating a spectrosome without displaying all of the
links.