feat: Allow to reuse the coordinates of a graph
The function spectrosome_chart now returns the coordinates of the
graphs it plots. These coordinates can be used in another call of the
function to plot the vertices at the exact same coordinates.
This can be useful to plot the same graph with another value for
another parameter like the one used for hidding links.