feat: Remove the ID columns from link attributes
Remove the ID columns from the two attributes which refer to the links
between nodes and the ones between patterns.
Change the order of the columns of the return data frame of function
spectrosome_chart, and therefore of function cluster_chart, to put the
ID first. As a result, the vertices and edges return data frames are
consistent.
This also has the consequence of reducing the final weight of an object
of class SpectralAnalyzer, especially if it has a lot of links.