feat: Fix the colors used for categories independently
The colors used to represent categories in spectrosomes and trees are
now independent of the number of values represented on the chart.
The colors are determined when the object is created and can be changed
by the user. Before that, they were defined according to the number of
values represented on a chart and were not consistent between different
charts.
The colors are stored in a new attribute: a list of vectors (one vector
for each category; values of the vector are the colors, named by the
values of the category).