docs: Add and complete the list of help pages
Add the functions relating to the MCR approach to the list of functions
of the spectral analysis example notebook.
Add one specific vignette to list the datasets, functions, classes and
methods in a organized way as the "list of help pages".
Add information to access the new vignette in the package help page and
in the readme file.
Remove the internal keyword so the package help page appear in the
index.