refactor: Simplify the use of SpectralAnalyzer objects
Merge the parameters "entities" and "characteristics" in all related
functions.
Rename parameters "nodes_characteristics" and
"patterns_characteristics" for consistency.
Add 2 functions to detect the type of entities stored in a data frame
and to convert shortcuts for nodes and patterns into data frame.
Add class attributes for the use of these 2 functions.
Update the "SA_instance" example object.
Change the place of the function "check_access_to_category" in the
file.