fix: Add error messages when dealing with THQ
Add error messages when dealing with THQ for unnamed values.
Creation of one function to check if the elements of a data structure
are named.
Include utils.R in mixer.R to be able to use this function.
In all functions dealing with THQ, check if the given values are named
and stop if they are not, except in the one computing THQ.