feat: Allow to extract subset of values from a class
Export the function subset_from_class.
Allow to extract a subset of values without giving the parameter
references.
Allow to use a list of classes and not just a logical matrix.
Add examples in the related documentation.

Update the function used to check the data for the approach using
classes: allows to check the values only and not necessarily values and
references.