feat: Improve spectrosome visualization
Change the symbols used for links in the legend (circles became lines).
Correct the symbol size setting.
Reduce the size of the bigger symbol in the legend.
Fix the position of the parts of the legend (especially the numbers
relating to status).
Correct the size of the legend box so that the graph do not overlap the
legend.
Move slightly the title and the subtitle and reduce their region.
Change title sizes for consistency with the spectrum and itemset chart.