feat: Add a class ObservationSet
Add a class ObservationSet for representing observations replacing
variables of type list containing the observations.
Add functions for subsetting and listing itemsets.
Replace the function turn_obs_into_transactions by the function as
with the classes ObservationSet and transactions.

Rename make_INRS_observations to make_OE_observations.
Add parameter keys to make_OE_observations and its sub-functions so as
not to constrain the existence of variables named ID, CODE and YEAR.
Add an alternative to the constraint of the existence of a variable
named WS.ID for the definition of work situations: use of the first
column if no one is named WS.ID.
The functions make_observations and make_OE_observations now return
ObservationSet objects.

Correct the way the SpectralAnalyzer class and validation function are
defined.
Update the use of observations in the class SpectralAnalyzer, in
examples and documentation.
The tutorial about SpectralAnalyzer is outdated and will be fully
revised soon.