feat: Allow to plot itemset chart from ObservationSet
Allow to plot the itemsets from the observations using an object of
class SpectralAnalyzer.
Add one sub-function for preparing data from the observations and call
it both in method for signature ObservationSet and in method for
signature SpectralAnalyzer.
Change the sort of items in itemset_chart for signature
SpectralAnalyzer: if sort_by is category, the items are sorted according
to the category then to there names if use_names and has_item_names are
TRUE; according to the category then to there identification codes
otherwise.