feat: Change item sorting in co_occurrence_chart
Change the sort of items in co_occurrence_chart for signature
SpectralAnalyzer. If a category is used, they are sorted according to
the category then to there names if use_names and has_item_names are
TRUE; according to the category then to their identification codes if
they are FALSE. (Instead of category and identification codes for both
possibility).