feat: Enhance the itemset chart
Fix the size of the area in which the sizes of the itemsets are
displayed, regardless of the chart.
Simplify redefinitions of margins and coordinates.
Correct the offset of the margin associated with the items.
Replace semicircles signifying items by circles.
Adapt the graphics area to the text displayed over the itemsets (in
addition to the text displayed under them).

Remaining issues (although the code looks correct):
 - The left margin seems to be different from the right one (cf. title
   position and line limit).
 - Position of the item labels is modified if the offset is modified.
   It is also the case if the initial margin is changed.

Note: there is a bug in the operation of RStudioGD concerning the
computation of string width when there are accented characters.
This results in a different position of the item labels depending on
the text. Only on this graphics device.