feat: Add one information in the MCR summary
Add one column in the data frames returned by the functions mcr_summary,
mcr_summary_for_list and mcr_summary_by_class. This column contains the
number of values of each set of values.
Useful because some indicators are interpreted according to this number.
Update the related vignette and manuals.