fix: Correct issues related to transaction creation
Several issues fixed about the function make_transactions for creating
lists of transactions or TransactionSet objects.

1. Internal call to the function by() to aggregate data was too long to
execute and generated memory overflows when large data fram was used
and the aggregation was made using several variables. This was because
of the application of the function unique (i.e., to only keep unique
values after the aggregation). The aggregation is now always done with
the application of the function c, and the remove of duplicate values
is made afterwards.

2. Transactions were somes list of vectors instead of list of lists
especially when aggregated data contain only one value for each
variable. Moreover, this resulted in an error if the expected output
was a TransactionSet object (i.e., if the parameter item_key is given).
The way the transactions are converted to list before returning them (or
before returning the related TransactionSet object) has been changed to
avoid that.

3. Factors sometimes converted to numeric when the parameter
unique_values was different from TRUE, because of the use of the
function c for the aggregation. Factors are now converted to characters
before the aggregation instead of after.

By the way, start to also comment in English inside functions.