chore: Improve gitignore
Adjust the content of the gitignore file to what
devtools::build_vignettes() is setting while building vignettes.
The patterns proposed by this function could not be used before
because it matched not only the ./doc folder but also the ./inst/doc
one. Now it only match the former.
This means that changes to gitignore no longer have to be discarded each
time vignettes are built. Further vignette building will no longer
change the gitignore file.