feat: Improve THQ pairs contingency tables output
In function thq_pairs from the MCR approach:
when the argument alone is TRUE, the column and row added and named NULL
are now always at the end of the table (last column and last row).
Before that, these column and row were at the end only if there was no
element that was alone (and alone is TRUE) and were in the middle of the
table according to alphabetic order in the other cases.