perf: Reduce execution time for MCR summary by class
Function mcr_summary now returns a list of the main indicators instead
of a one-row data frame if the given values are a vector or a list of a
single vector. Creation of the data frames when computing MCR summary by
class is therefore faster.
Update function manuals of mcr_summary and mcr_summary_for_list
accordingly.
Use of the function which.max instead of sort when finding the first top
hazard quotient in function top_hazard_quotient (which is called once on
each subset of values).
Examples of time improvement for two instances (before -> now):
 - 30,000 sets of values:  01m 56s ->     26s
 - 184,580 sets of values: 13m 44s -> 02m 58s