feat: Allow to compute and plot cooccurrence proportions
Add a parameter named proportions to the functions co_occurrence_matrix
and co_occurrence_chart. It allows to compute or plot co-occurrence
proportions instead of co-occurrences.
Add two parameters to the internal function plot_heb_chart to
parametrize the main scale name and its breaks.
Add examples to the function manuals.
Add examples and description to the vignettes about TransactionSet and
TransactionAnalyzer.