docs: Change format of files underlying PDFdocumentation
Change the format of the raw files used to generate the PDF
documentation (i.e., exported files describing the attributes and the
methods of the classes TransactionSet and TransactionAnalyzer).
They were DOCX files; they are now ODT files.
New ODT files are heavier but they are not installed with the package
whereas PDF files are.
New versions of the PDF files are lighter: total of 200 KB instead of
560 KB previously.