fix: Correct the diagonal of the co-occurrences matrix
Previously, for the diagonal of the matrix, the function
co_occurrence_matrix counted the number of occurrences of each item in
the transactions considering duplicates items in the transactions. Now,
as described in the manual, it counts the number of transactions
containing each item. It is therefore different from calling the
function table_on_list if duplicates exist.