fix: Correct names of hazard quotient values
In the function hazard_quotient, names were given to the returned hazard
quotients if the given values were a matrix of sets of values. These
names were V1, V2, ... They now correspond to the names of the sets of
values in the matrix given as argument. They can also be NULL.
This affects other functions, like hazard_index: the returned values are
now named the same way if the argument values or the argument hq is
given. Before that, returned values were named V1, V2... if given values
were a matrix but named according to the hazard quotient if hq was given
as a matrix.