feat: Change mcr_summary return to data frame if list
The function computing almost all indicators of the MCR approach
(mcr_summary) previously returned a list if the given values were a
vector or a list of a single vector, and a data frame if the given
values were a matrix or a list of several vectors.
It now returns a list only if the given values are a vector, and a data
frame otherwise.