feat: Change patterns per year matrix type to integer
Change the type of the data contained in the attribute patterns_per_year
of objects of class TransactionAnalyzer. It was numeric. It is now
integer. This leads to a reduction of the size of the resulting object
and a homogenization since the related attribute nodes_per_year is of
type integer.

Update the dataset TA_instance accordingly.