fix: Correct minimum support issue when mining itemsets
The workaround used was not enough. The issue with the function eclat
from arules was that no frequent itemsets were found when all had
exactly a support equal to the parameter defining the minimum support to
consider an itemset as frequent.
The bug is fixed in the current development version of arules on GitHub
(1.7-3.1) and will be added to the next release version on CRAN.
Remove the workaround previously added in the method
list_separate_patterns of the class TransactionAnalyzer.