fix: Correct specific warning and error in mcr_chart
When calling function mcr_chart, fix warning if parameter regions is
FALSE and error if it is TRUE in specific conditions.
The conditions to trigger these were that only the region corresponding
to the group II contained points to plot and the limit of the plot did
not contained the limits of the regions.