fix: Prevent the use of comma character in item codes
Add a test in the initializer of the class TransactionAnalyzer to
prevent the use of comma in the given item codes.
Allowing this character implies to change each call of the functions
vector_notation and set_notation; to add them a parameter to specify
the separator character; but more importanly, add parameters in the
functions extract_rules and rules_chart to specify which separator to
use in the returned data or that is used in the given data. The last
point would add complexity and the need to explain why it is necessary
to use another character than the comma as separator in specific cases.