feat: Function reduce_sets now always reduces references
In function reduce_sets, argument references is now processed if it is
not a list. This is needed if duplicated names exist in the values when
they are a vector or a matrix, since there are as many references as
there are values names.
Update manual documentation and add unit tests.