feat: Allow to plot new indicators about rules
Allow to plot the specificity, accuracy or added value using the method
rules_chart.
Add a new validation of the presence of the desired indicator in the
given data.frame.
Change the call to extract_rules if no rules are given and the desired
indicator is one of the three new ones.
Consider divergent palettes for the indicateor named added value since
it can be lower than 0.
Change the default value of the argument threshold from 0 to NULL since
one of the new indicators can be lower than 0.