feat: Remove one set of color palettes in rules charts
In the method rules_chart, the set of palettes used for the
characteristics confidence and specificity are now the same as those
for the other characteristics (except the added value).
In the DESCRIPTION file, move the package name RColorBrewer from the
Suggests section to the Imports section. So it has been added to the
set of packages that need to be installed in the README installation
section.