docs: Improve the DESCRIPTION file
In the DESCRIPTION file:
 - Correct the field Author@R to Authors@R.
 - Remove fields Author and Maintainer since they are therefore no
   longer needed.
 - Add field URL and BugReports.
 - Reorder fields.
As a consequence, the main package help page now contains the links to
the github page and to the issues page.
Update the file CITATION to consider the new DESCRIPTION field URL and
to consider the authors excluding the last one (which is currently the
organization being the copyright holder and funder) since it does not
have the actual author role.