fix: Correct the function thq_pairs
Correct the function thq_pairs_for_list according to the last changes of
the function top_hazard_quotient.
Since HQ values equal to 0 are renamed to NA in top_hazard_quotient,
added NULL values in order to have pairs when the argument alone is
TRUE, can no longer be values equal to 0. They are now values equal to
1 (as it is in thq_pairs_for_matrix).