feat: Add specific cases and results to MCR indicators
Change the functions maximum_cumulative_ratio, missed_toxicity and
reciprocal_of_mcr to consider the specific case in which all values are
equal to 0. Instead of returning NaN (because the calculation cannot be
done), these functions now return 0.
Update help pages related to these indicators and update the vignette to
describe this specific case.
Vignette has not been re-generated.