fix: Correct the creation of transaction names
Data frames are coerced to matrices when applying the apply function.
This means that integer columns are coerced to character not by the
paste function but by the apply function. Moreover, coercing a data
frame to a matrix implies the use of the format function with argument
trim = FALSE. This results in leading whitespaces so that all integer
values are coerced to character values having the same number of
characters.
The use of the apply function has been removed for the use of function
do.call when creating transaction names in make_transactions.
This new way avoid creating leading whitespaces and is faster.