fix: Correct abnormal execution time when timing is on
Add an argument gc to function display_time used when calling function
system.time. It is used to choose whether to call the garbage collector
before timing.
The default value of this new argument is FALSE, whereas previous states
of the function used TRUE in the call to system.time.
This caused abnormally long execution time because of calls to the GC,
especially when the memory usage was high. This also caused incorrect
results in timing, because calls to GC were not included during timing.