fix: Correct the initialization of nodes in TA objects
In method list_separate_trx of class TransactionAnalyzer, change the
regular expression used to create nodes according to row names of the
data frame containing node frequencies for each year.
When the total number of characters of a row name exceed a certain
threshold and this row name is based on a vector, the new line character
is automatically added by R.
This led to an error when initializing TransactionAnalyzer objects
having nodes containing lot of items or long item identifiers.